requirements.txt
# Ingestion
entsoe-py>=0.6.0
openmeteo-requests>=1.1.0
requests-cache>=1.1.0
retry-requests>=2.0.0
holidays>=0.45

# Core
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=14.0.0
duckdb>=1.2.0

# Feature store
feast>=0.40.0

# ML
scikit-learn>=1.5.0
lightgbm>=4.3.0
joblib>=1.3.0

# MLOps / pipeline orchestration
zenml[server]>=0.64.0

# Serving
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0

# Dashboard
streamlit>=1.40.0
plotly>=5.22.0

# Utils
python-dotenv>=1.1.0
tqdm>=4.67.0
jupyterlab>=4.2.0
ipykernel>=6.29.0
